feat: 비회원 처리 로직 추가

+++ src/main/java/nextstep/subway/favorite/ui/FavoriteController.java
@@ -2,10 +2,12 @@ package nextstep.subway.favorite.ui;
 
 import nextstep.subway.auth.domain.AuthenticationPrincipal;
 import nextstep.subway.auth.domain.LoginMember;
+import nextstep.subway.auth.domain.Member;
 import nextstep.subway.favorite.application.FavoriteService;
 import nextstep.subway.favorite.domain.Favorite;
 import nextstep.subway.favorite.dto.FavoriteRequest;
 import nextstep.subway.favorite.dto.FavoriteResponse;
+import nextstep.subway.member.dto.MemberResponse;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -22,19 +24,25 @@ public class FavoriteController {
     }
 
     @PostMapping("/favorites")
-    public ResponseEntity createFavorite(@AuthenticationPrincipal LoginMember loginMember, @RequestBody FavoriteRequest favoriteRequest) {
+    public ResponseEntity createFavorite(@AuthenticationPrincipal Member loginMember, @RequestBody FavoriteRequest favoriteRequest) {
         Favorite favorite = favoriteService.createFavorite(loginMember.getId(), favoriteRequest);
         return ResponseEntity.created(URI.create("/favorites/" + favorite.getId())).body(favorite);
     }
 
     @GetMapping("/favorites")
-    public ResponseEntity findAllFavorites(@AuthenticationPrincipal LoginMember loginMember) {
+    public ResponseEntity findAllFavorites(@AuthenticationPrincipal Member loginMember) {
+        if (!loginMember.isLoginMember()) {
+            return ResponseEntity.ok().body(MemberResponse.nonLoginMemberResponse());
+        }
         List<FavoriteResponse> favoriteResponses = favoriteService.findAllFavorites(loginMember.getId());
         return ResponseEntity.ok(favoriteResponses);
     }
 
     @DeleteMapping("/favorites/{favoriteId}")
-    public ResponseEntity deleteFavorite(@AuthenticationPrincipal LoginMember loginMember, @PathVariable Long favoriteId) {
+    public ResponseEntity deleteFavorite(@AuthenticationPrincipal Member loginMember, @PathVariable Long favoriteId) {
+        if (!loginMember.isLoginMember()) {
+            return ResponseEntity.ok().body(MemberResponse.nonLoginMemberResponse());
+        }
         favoriteService.deleteFavorite(favoriteId);
         return ResponseEntity.noContent().build();
     }
