+++ src/main/java/nextstep/subway/member/ui/MemberController.java
@@ -3,6 +3,7 @@ package nextstep.subway.member.ui;
 import nextstep.subway.auth.domain.LoginMember;
 import nextstep.subway.auth.domain.AuthenticationPrincipal;
 import nextstep.subway.auth.domain.LoginMember;
+import nextstep.subway.auth.domain.Member;
 import nextstep.subway.auth.exception.InValidTokenException;
 import nextstep.subway.member.application.MemberService;
 import nextstep.subway.member.dto.MemberRequest;
@@ -47,19 +48,28 @@ public class MemberController {
     }
 
     @GetMapping("/members/me")
-    public ResponseEntity<MemberResponse> findMemberOfMine(@AuthenticationPrincipal LoginMember loginMember) {
+    public ResponseEntity<MemberResponse> findMemberOfMine(@AuthenticationPrincipal Member loginMember) {
+        if (!loginMember.isLoginMember()) {
+            return ResponseEntity.ok().body(MemberResponse.nonLoginMemberResponse());
+        }
         MemberResponse member = memberService.findMember(loginMember.getId());
         return ResponseEntity.ok().body(member);
     }
 
     @PutMapping("/members/me")
-    public ResponseEntity<MemberResponse> updateMemberOfMine(@AuthenticationPrincipal LoginMember loginMember, @RequestBody MemberRequest param) {
+    public ResponseEntity<MemberResponse> updateMemberOfMine(@AuthenticationPrincipal Member loginMember, @RequestBody MemberRequest param) {
+        if (!loginMember.isLoginMember()) {
+            return ResponseEntity.ok().body(MemberResponse.nonLoginMemberResponse());
+        }
         memberService.updateMember(loginMember.getId(), param);
         return ResponseEntity.ok().build();
     }
 
     @DeleteMapping("/members/me")
-    public ResponseEntity<MemberResponse> deleteMemberOfMine(@AuthenticationPrincipal LoginMember loginMember) {
+    public ResponseEntity<MemberResponse> deleteMemberOfMine(@AuthenticationPrincipal Member loginMember) {
+        if (!loginMember.isLoginMember()) {
+            return ResponseEntity.ok().body(MemberResponse.nonLoginMemberResponse());
+        }
         memberService.deleteMember(loginMember.getId());
         return ResponseEntity.noContent().build();
     }
